+++ Painel/painel/Painel.c
@@ -12,7 +12,7 @@
 
 typedef struct {
     TCHAR name[TAM_USERNAME];
-    DWORD points;
+    float points;
 }PLAYER_MP;
 typedef struct {
     DWORD num_letras;
